requirements.txt
Django~=5.2.1
dotenv~=0.9.9
python-dotenv~=1.1.0
djangorestframework~=3.16.0